Foram feitas algumas melhorias no código. Principalmente a identificação das funções repidas, que agora estão no arquivo utils.js.
