+++ src/components/FeaturedPokemon/index.js
@@ -4,65 +4,14 @@ import PokeTypes from "../../components/PokeTypes";
 import PokeStats from "../../components/PokeStats";
 import imgPokeball from "../../assets/pokeball.png";
 import _get from "lodash/get";
-import { colorTypeGradients } from "../../utils/utils";
-import { Link } from "react-router-dom";
+import {
+    colorTypeGradients,
+    splitName,
+    spriteAdapterOfficial,
+    zeroLeft,
+} from "../../utils/utils";
 
 function FeaturedPokemon(props) {
-    //Verrificando o tamanho do NOME do Pokemon, se preciso o nome será dividio e mostrado só o primeiro nome.
-    function splitName(name) {
-        let newName = "";
-        if (name.length > 15) {
-            let splitedName = name.split("-");
-            newName = splitedName[0];
-        } else {
-            newName = name.replace(/-/g, " ");
-        }
-        return newName;
-    }
-
-    //Pegando a URL da imagem oficial do Pokemon
-    function spriteAdapterOfficial(spriteOfficial) {
-        let oficial_atwork = _get(
-            spriteOfficial,
-            "other.official-artwork.front_default",
-            ""
-        );
-        let dream_word = _get(
-            spriteOfficial,
-            "other.dream_world.front_default",
-            ""
-        );
-
-        if (dream_word) {
-            return dream_word;
-        }
-
-        if (oficial_atwork) {
-            return oficial_atwork;
-        }
-
-        return null;
-    }
-
-    //Adicionando zero a esqueda no númeoro do Pokemon.
-    function zeroLeft(pokeId) {
-        if (pokeId < 10) {
-            return "000" + pokeId;
-        }
-
-        if (pokeId >= 10 && pokeId < 100) {
-            return "00" + pokeId;
-        }
-
-        if (pokeId >= 100 && pokeId < 1000) {
-            return "0" + pokeId;
-        }
-
-        if (pokeId >= 1000) {
-            return pokeId;
-        }
-    }
-
     const spriteOfficial = spriteAdapterOfficial(props.pokemon.sprites);
 
     //Definindo a cor do card com base na cor do tipo
