+++ src/components/PokemonTable/ColumnsOptions/index.js
@@ -1,5 +1,4 @@
 import React from "react";
-//import Checkbox from "../CheckBox";
 import * as S from "./styles";
 
 const ColumnsOptions = (props) => {
