+++ src/views/allPokemon/index.js
@@ -36,8 +36,6 @@ function AllPokemon() {
                     a.id > b.id ? 1 : b.id > a.id ? -1 : 0
                 );
 
-                //console.log(newPokeList);
-
                 setAllPokemon(newPokeList);
                 setRemoveLoading(true);
             }
